# Use `Float32` for the matrix coefficients for low precision colors