This also changes the definition of the sRGB transform matrix slightly. (cf. https://github.com/JuliaGraphics/Colors.jl/issues/355#issuecomment-539524932)

cf. #477