@kimikage 2021-06-20
### Benchmark
```julia
using Colors, FixedPointNumbers, BenchmarkTools
N = 1000;
rgb_f64 = rand(RGB{Float64}, N, N);
rgb_f32 = rand(RGB{Float32}, N, N);
rgb_n0f8 = rand(RGB{N0f8}, N, N);
xyz_f64 = XYZ.(rgb_f64);
xyz_f32 = XYZ.(rgb_f32);
yiq_f64 = YIQ.(rgb_f64);
yiq_f32 = YIQ.(rgb_f32);

@btime convert.(XYZ, $rgb_f64);
@btime convert.(XYZ, $rgb_f32);
@btime convert.(XYZ, $rgb_n0f8);
@btime convert.(RGB, $xyz_f64);
@btime convert.(RGB, $xyz_f32);
@btime convert.(RGB{N0f8}, $xyz_f32);

@btime convert.(YIQ, $rgb_f64);
@btime convert.(YIQ, $rgb_f32);
@btime convert.(YIQ, $rgb_n0f8);
@btime convert.(RGB, $yiq_f64);
@btime convert.(RGB, $yiq_f32);
@btime convert.(RGB{N0f8}, $yiq_f32);
```

|                               | before    | after     |
|:------------------------------|----------:|----------:|
|`RGB{Float64}`-->`XYZ{Float64}`| 30.209 ms | 30.318 ms |
|`RGB{Float32}`-->`XYZ{Float32}`| 25.528 ms | 23.325 ms |
|`RGB{N0f8}`   -->`XYZ{Float32}`|  3.506 ms |  2.396 ms |
|`XYZ{Float64}`-->`RGB{Float64}`| 36.156 ms | 34.093 ms |
|`XYZ{Float32}`-->`RGB{Float32}`| 47.960 ms | 36.715 ms |
|`XYZ{Float32}`-->`RGB{N0f8}`   | 47.224 ms | 42.136 ms |

|                               | before    | after     |
|:------------------------------|----------:|----------:|
|`RGB{Float64}`-->`YIQ{Float64}`|  5.379 ms |  5.245 ms |
|`RGB{Float32}`-->`YIQ{Float32}`|  3.076 ms |  2.373 ms |
|`RGB{N0f8}`   -->`YIQ{Float32}`|  9.060 ms |  2.006 ms |
|`YIQ{Float64}`-->`RGB{Float64}`|  5.709 ms |  5.755 ms |
|`YIQ{Float32}`-->`RGB{Float32}`|  3.398 ms |  2.546 ms |
|`YIQ{Float32}`-->`RGB{N0f8}`   | 18.193 ms | 16.755 ms |