@kimikage 2021-06-22
Now that the `RGB`<-->`XYZ` conversions avoids the precompilation problem, I think we are ready to merge this.